refactor(core): extract shared ParseEuropeanDecimal into CsvParsing
